pyproject.toml
[project]
name = "sediment"
version = "0.1.0"
description = "Tacit knowledge extraction system for AI Agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0.2",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "starlette>=0.37.2",
    "uvicorn>=0.30.0",
]

[project.scripts]
sediment = "sediment.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "playwright>=1.52.0",
    "ruff>=0.4.0",
    "pytest>=8.0.0",
]

[tool.hatch.build]
include = [
    "src/sediment/**/*.py",
    "src/sediment/**/*.html",
    "src/sediment/**/*.js",
    "src/sediment/**/*.md",
    "src/sediment/**/*.yaml",
    "src/sediment/**/*.svg",
    "README.md",
    "README_zh.md",
    "LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["src/sediment"]

[tool.hatch.build.targets.sdist]
include = [
    "src",
    "tests",
    "scripts",
    "design",
    "benchmarks/TEST_PLAN.md",
    "benchmarks/scripts/*.py",
    "README.md",
    "README_zh.md",
    "LICENSE",
    "pyproject.toml",
    "uv.lock",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]
extend-exclude = ["benchmarks"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "BLE"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: deterministic single-surface tests with no live server",
    "integration: cross-module tests using shared in-process fixtures",
    "e2e: scenario-driven end-to-end workflow coverage",
    "browser: browser-backed E2E coverage",
]
